chicago manual of style says...
lease see CMOS 14.21: “A note number should generally be placed at the
end of a sentence or at the end of a clause. The number normally
follows a quotation (whether it is run into the text or set as an
extract). Relative to other punctuation, the number follows any
punctuation mark except for the dash, which it precedes.” See 14.21 for
examples.